+++ pkg/compactor/default_compactor.go
@@ -15,27 +15,26 @@ import (
 	"github.com/grafana/mimir/pkg/ring"
 )
 
-var (
-	DefaultBlocksGrouperFactory = func(ctx context.Context, cfg Config, cfgProvider ConfigProvider, bkt objstore.Bucket, userID string, ring *ring.Ring, instanceAddr string, logger log.Logger, reg prometheus.Registerer) Grouper {
-		return NewDefaultGrouper(
-			logger,
-			bkt,
-			false, // Do not accept malformed indexes
-			metadata.NoneFunc)
+func defaultBlocksGrouperFactory(ctx context.Context, cfg Config, cfgProvider ConfigProvider, bkt objstore.Bucket, userID string, ring *ring.Ring, instanceAddr string, logger log.Logger, reg prometheus.Registerer) Grouper {
+	return NewDefaultGrouper(
+		userID,
+		logger,
+		bkt,
+		false, // Do not accept malformed indexes
+		metadata.NoneFunc)
+}
+
+func defaultBlocksCompactorFactory(ctx context.Context, cfg Config, logger log.Logger, reg prometheus.Registerer) (Compactor, Planner, error) {
+	compactor, err := tsdb.NewLeveledCompactor(ctx, reg, logger, cfg.BlockRanges.ToMilliseconds(), downsample.NewPool(), nil)
+	if err != nil {
+		return nil, nil, err
 	}
 
-	DefaultBlocksCompactorFactory = func(ctx context.Context, cfg Config, logger log.Logger, reg prometheus.Registerer) (Compactor, Planner, error) {
-		compactor, err := tsdb.NewLeveledCompactor(ctx, reg, logger, cfg.BlockRanges.ToMilliseconds(), downsample.NewPool(), nil)
-		if err != nil {
-			return nil, nil, err
-		}
+	planner := NewTSDBBasedPlanner(logger, cfg.BlockRanges.ToMilliseconds())
+	return compactor, planner, nil
+}
 
-		planner := NewTSDBBasedPlanner(logger, cfg.BlockRanges.ToMilliseconds())
-		return compactor, planner, nil
-	}
-)
-
-func ConfigureDefaultCompactor(cfg *Config) {
-	cfg.BlocksGrouperFactory = DefaultBlocksGrouperFactory
-	cfg.BlocksCompactorFactory = DefaultBlocksCompactorFactory
+func configureDefaultCompactor(cfg *Config) {
+	cfg.BlocksGrouperFactory = defaultBlocksGrouperFactory
+	cfg.BlocksCompactorFactory = defaultBlocksCompactorFactory
 }